+++ src/resolvers/versionedSection.ts
@@ -4,12 +4,13 @@ import { VersionedSection, VersionedSectionSearchResult } from "../models/Versio
 import { Section } from "../models/Section";
 import { Tag } from "../models/Tag";
 import { VersionedTemplate } from "../models/VersionedTemplate";
-import { InternalServerError } from "../utils/graphQLErrors";
+import { AuthenticationError, ForbiddenError, InternalServerError } from "../utils/graphQLErrors";
 import { VersionedQuestion } from "../models/VersionedQuestion";
 import { prepareObjectForLogs } from "../logger";
 import { GraphQLError } from "graphql";
 import { PaginationOptionsForCursors, PaginationOptionsForOffsets, PaginationType } from "../types/general";
 import { isNullOrUndefined, normaliseDateTime } from "../utils/helpers";
+import { isAuthorized } from "../services/authService";
 
 export const resolvers: Resolvers = {
   Query: {
@@ -43,6 +44,21 @@ export const resolvers: Resolvers = {
         throw InternalServerError();
       }
     },
+    // Get a specific VersionedSection
+    publishedSection: async (_, { versionedSectionId }, context: MyContext): Promise<VersionedSection> => {
+      const reference = 'publishedSection resolver';
+      try {
+        if (isAuthorized(context.token)) {
+          return await VersionedSection.findById(reference, context, versionedSectionId);
+        }
+        throw context?.token ? ForbiddenError() : AuthenticationError();
+      } catch (err) {
+        if (err instanceof GraphQLError) throw err;
+
+        context.logger.error(prepareObjectForLogs(err), `Failure in ${reference}`);
+        throw InternalServerError();
+      }
+    }
   },
 
   VersionedSection: {